Part 1 - refactor utf8 utilities to shared
